+++ test/audits/forms.test.js
@@ -118,6 +118,9 @@ describe('opportunities handler method', () => {
       env: {
         S3_SCRAPER_BUCKET_NAME: 'test-bucket',
       },
+      site: {
+        getId: sinon.stub().returns('test-site-id'),
+      },
     };
     auditData = testData.auditData3;
   });
@@ -377,6 +380,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       save: sinon.stub(),
       getType: () => 'high-page-views-low-form-nav',
       setData: sinon.stub(),
+      setGuidance: sinon.stub(),
       getData: sinon.stub().returns({
         form: 'https://www.surest.com/newsletter',
         screenshot: '',
@@ -404,6 +408,9 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       env: {
         S3_SCRAPER_BUCKET_NAME: 'test-bucket',
       },
+      site: {
+        getId: sinon.stub().returns('test-site-id'),
+      },
     };
     auditData = testData.oppty2AuditData;
   });
@@ -428,10 +435,11 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
         formViews: 300,
         pageViews: 8670,
         samples: 8670,
+        scrapedStatus: false,
         metrics: [
           {
             type: 'conversionRate',
-            vendor: '*',
+            device: '*',
             value: {
               page: null,
             },
@@ -441,6 +449,16 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
           source: '#teaser-related02 .cmp-teaser__action-link',
           url: 'https://www.surest.com/about-us',
         },
+      },
+      guidance: {
+        recommendations: [
+          {
+            insight: 'The CTA element in the page: https://www.surest.com/about-us is not placed in the most optimal positions for visibility and engagement',
+            recommendation: 'Reposition the CTA to be more centrally located and ensure they are above the fold.',
+            type: 'guidance',
+            rationale: 'CTAs placed above the fold and in central positions are more likely to be seen and clicked by users, leading to higher engagement rates.',
+          },
+        ],
       },
     };